streamlit-job-app/src/utils/groq_scraper.py
import os
import requests
from bs4 import BeautifulSoup

def scrape_recruiter_email(company_name):
    # Example: search for recruiter email using company name
    search_url = f"https://www.google.com/search?q={company_name}+recruiter+email"
    headers = {"User-Agent": "Mozilla/5.0"}
    response = requests.get(search_url, headers=headers)
    soup = BeautifulSoup(response.text, "html.parser")
    # This is a placeholder. You need to implement actual scraping logic.
    # For now, just return None.
    return None

def get_email_from_groq(company_name):
    # Placeholder for Groq API integration
    # This function should use the Groq client to scrape for emails
    pass  # Implement Groq API call here if needed