streamlit-job-app/requirements.txt
streamlit
pandas
openpyxl
python-docx
PyPDF2
requests
dotenv
groq
beautifulsoup4